+++ test/predictedspeeds.cc
@@ -1,7 +1,3 @@
-#include <boost/archive/iterators/base64_from_binary.hpp>
-#include <boost/archive/iterators/binary_from_base64.hpp>
-#include <boost/archive/iterators/transform_width.hpp>
-
 #include <iostream>
 
 #include "baldr/predictedspeeds.h"
@@ -14,13 +10,6 @@ using namespace valhalla::baldr;
 using namespace valhalla::midgard;
 
 namespace {
-
-// base64 decoding
-std::string decode64(const std::string& val) {
-  using namespace boost::archive::iterators;
-  using It = transform_width<binary_from_base64<std::string::const_iterator>, 8, 6>;
-  return std::string(It(std::begin(val)), It(std::end(val)));
-}
 
 // Convert big endian bytes to little endian
 int16_t to_little_endian(const int16_t val) {
@@ -63,7 +52,12 @@ TEST(PredicteSpeeds, test_decoding) {
 
   // Decode the base64 string and cast the data to a raw string of signed bytes
   auto decoded_str = decode64(encoded_speed_string);
-  EXPECT_EQ(decoded_str.size(), 402);
+  // HACK(mookerji): kDecodedSpeedSize+1 is the expected size (as opposed to kDecodedSpeedSize)
+  // because we start reading from a 1-byte offset below at the little endian conversion. This is
+  // actually a broken test fixture because we start reading from 0 in the actual CLI decoding in
+  // src/mjolnir/valhalla_add_predicted_traffic.cc. To FIX this, we need to encode the speeds[]
+  // array below and start the little endian conversion from 0 instead of 1.
+  EXPECT_EQ(decoded_str.size(), kDecodedSpeedSize + 1);
 
   auto raw = reinterpret_cast<const int8_t*>(decoded_str.data());
 
@@ -73,8 +67,8 @@ TEST(PredicteSpeeds, test_decoding) {
   // Create the coefficients. Each group of 2 bytes represents a signed, int16 number (big endian).
   // Convert to little endian.
   int idx = 1;
-  int16_t coefficients[200];
-  for (uint32_t i = 0; i < 200; ++i, idx += 2) {
+  int16_t coefficients[kCoefficientCount];
+  for (uint32_t i = 0; i < kCoefficientCount; ++i, idx += 2) {
     coefficients[i] = to_little_endian(*(reinterpret_cast<const int16_t*>(&raw[idx])));
   }
 
@@ -197,7 +191,8 @@ TEST(PredicteSpeeds, test_negative_speeds) {
 
   // Decode the base64 string and cast the data to a raw string of signed bytes
   auto decoded_str = decode64(encoded_speed_string);
-  EXPECT_EQ(decoded_str.size(), 402);
+  // HACK(mookerji): See note above.
+  EXPECT_EQ(decoded_str.size(), kDecodedSpeedSize + 1);
 
   auto raw = reinterpret_cast<const int8_t*>(decoded_str.data());
 
@@ -207,8 +202,8 @@ TEST(PredicteSpeeds, test_negative_speeds) {
   // Create the coefficients. Each group of 2 bytes represents a signed, int16 number (big endian).
   // Convert to little endian.
   int idx = 1;
-  int16_t coefficients[200];
-  for (uint32_t i = 0; i < 200; ++i, idx += 2) {
+  int16_t coefficients[kCoefficientCount];
+  for (uint32_t i = 0; i < kCoefficientCount; ++i, idx += 2) {
     coefficients[i] = to_little_endian(*(reinterpret_cast<const int16_t*>(&raw[idx])));
   }